requirements.txt
langchain
langchain-core
langchain-community
langchain-experimental
pymupdf
langchain-text-splitters
faiss-cpu
langchain-ollama
langchain-openai